CTS: fix deletion of the tmp root for augeas
git-svn-id: http://svn.fedorahosted.org/svn/corosync/trunk@2679 fd59a12c-fef9-0310-b244-a6a79926bd2f

+++ cts/corosync.py
@@ -60,7 +60,8 @@ class CoroConfig(object):
             self.corobase = corobase
         example = self.corobase + "/conf/corosync.conf.example"
 
-        shutil.rmtree (self.new_root)
+        if os.path.isdir(self.new_root):
+            shutil.rmtree (self.new_root)
         os.makedirs (self.new_root + "/etc/corosync")
         shutil.copy (example, self.new_root + "/etc/corosync/corosync.conf")
